@@ -3,34 +3,25 @@
 namespace App\Entity;
 
 use App\Repository\InstanceRepository;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity(repositoryClass=InstanceRepository::class)
- */
+#[ORM\Entity(repositoryClass: InstanceRepository::class)]
 class Instance {
 
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private ?int $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
+    private ?int $id = null;
 
-    /**
-     * @ORM\Column(type="string")
-     */
-    private string $name;
+    #[ORM\Column(type: Types::STRING)]
+    private ?string $name = null;
 
-    /**
-     * @ORM\Column(type="string")
-     */
-    private string $mode;
+    #[ORM\Column(type: Types::STRING)]
+    private ?string $mode = null;
 
-    /**
-     * @ORM\Column(type="string")
-     */
-    private string $code;
+    #[ORM\Column(type: Types::STRING)]
+    private ?string $code = null;
 
     public function getId(): ?int {
         return $this->id;
@@ -40,7 +31,7 @@ class Instance {
         return $this->name;
     }
 
-    public function setName(string $name): self {
+    public function setName(?string $name): self {
         $this->name = $name;
 
         return $this;
@@ -50,7 +41,7 @@ class Instance {
         return $this->mode;
     }
 
-    public function setMode(string $mode): self {
+    public function setMode(?string $mode): self {
         $this->mode = $mode;
 
         return $this;
@@ -60,7 +51,7 @@ class Instance {
         return $this->code;
     }
 
-    public function setCode(string $code): self {
+    public function setCode(?string $code): self {
         $this->code = $code;
 
         return $this;
