@@ -3,50 +3,47 @@
 namespace App\Entity;
 
 use App\Repository\UserRepository;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
-/**
- * @ORM\Entity(repositoryClass=UserRepository::class)
- */
+#[ORM\Entity(repositoryClass: UserRepository::class)]
 #[UniqueEntity(fields: ['email'], message: 'There is already an account with this email')]
-class User implements UserInterface, PasswordAuthenticatedUserInterface
-{
+class User implements UserInterface, PasswordAuthenticatedUserInterface {
+
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
+    private ?int $id = null;
+
+    #[ORM\Column(type: Types::STRING, length: 180, unique: true)]
+    private ?string $email = null;
+
+    #[ORM\Column(type: Types::JSON)]
+    private ?array $roles = [];
+
     /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
+     * The hashed password
      */
-    private $id;
-    /**
-     * @ORM\Column(type="string", length=180, unique=true)
-     */
-    private $email;
-    /**
-     * @ORM\Column(type="json")
-     */
-    private $roles = [];
-    /**
-     * @var string The hashed password
-     * @ORM\Column(type="string")
-     */
-    private $password;
-    public function getId(): ?int
-    {
+    #[ORM\Column(type: Types::STRING)]
+    private ?string $password = null;
+
+    public function getId(): ?int {
         return $this->id;
     }
-    public function getEmail(): ?string
-    {
+
+    public function getEmail(): ?string {
         return $this->email;
     }
-    public function setEmail(string $email): self
-    {
+
+    public function setEmail(?string $email): self {
         $this->email = $email;
 
         return $this;
     }
+
     /**
      * A visual identifier that represents this user.
      *
@@ -56,6 +53,7 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return (string) $this->email;
     }
+
     /**
      * @deprecated since Symfony 5.3, use getUserIdentifier instead
      */
@@ -63,6 +61,7 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return (string) $this->email;
     }
+
     /**
      * @see UserInterface
      */
@@ -74,12 +73,14 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
         return array_unique($roles);
     }
+
     public function setRoles(array $roles): self
     {
         $this->roles = $roles;
 
         return $this;
     }
+
     /**
      * @see PasswordAuthenticatedUserInterface
      */
@@ -87,12 +88,14 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return $this->password;
     }
+
     public function setPassword(string $password): self
     {
         $this->password = $password;
 
         return $this;
     }
+
     /**
      * Returning a salt is only needed, if you are not using a modern
      * hashing algorithm (e.g. bcrypt or sodium) in your security.yaml.
@@ -103,10 +106,11 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     {
         return null;
     }
+
     /**
      * @see UserInterface
      */
-    public function eraseCredentials()
+    public function eraseCredentials(): void
     {
         // If you store any temporary, sensitive data on the user, clear it here
         // $this->plainPassword = null;
