@@ -13,8 +13,7 @@ use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 class LoginController extends AbstractController
 {
     #[Route('/login', name: 'app_login')]
-    public function index(AuthenticationUtils $authenticationUtils): Response
-    {
+    public function index(AuthenticationUtils $authenticationUtils): Response {
         // get the login error if there is one
         $error = $authenticationUtils->getLastAuthenticationError();
         // last username entered by the user
@@ -27,8 +26,7 @@ class LoginController extends AbstractController
     }
 
     #[Route('/addRight', name: 'add_right', methods: ['POST'])]
-    public function addRight(EntityManagerInterface $manager, Request $request): Response
-    {
+    public function addRight(EntityManagerInterface $manager, Request $request): Response {
         $email = $request->request->get('email');
         $user = $manager->getRepository(User::class)->findOneBy(['email' => $email]);
         $user?->setRoles(['ROLE_ADMIN']);
